+++ lib/clixon/clixon_xml_map.h
@@ -64,7 +64,8 @@ int xml_sanity(cxobj *x, void  *arg);
 int xml_non_config_data(cxobj *xt, void *arg);
 
 int xml2xpath(cxobj *x, char **xpath);
-int assign_namespaces(cxobj *x0, cxobj *x1, cxobj *x1p);
+int assign_namespace_element(cxobj *x0, cxobj *x1, cxobj *x1p);
+int assign_namespace_body(cxobj *x0, char *x0bstr, cxobj *x1);
 int xml_merge(cxobj *x0, cxobj *x1, yang_stmt *yspec, char **reason);
 int yang_enum_int_value(cxobj *node, int32_t *val);